Inventory button GUI working
Now to make them actually do something
Also to construct the interior chaimber for the scroll bar which will
hold the list of crew and their items.
First building horiz container that goes crew name (maybe icon) weapon
armor tool.
Then making a thing that has a column of those and making that the
scrollbar's content.

I also need to probably make the scrollbar check for the size of its
content and update accordingly as it does for top level items right now.